requirements.txt
altgraph==0.17.4
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
bottle==0.12.19
bottle-websocket==0.2.9
by==0.0.4
certifi==2023.7.22
cffi==1.15.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
Django==4.2.7
djangorestframework==3.14.0
Eel==0.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
Flask==3.0.0
fonttools==4.40.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.0.2
parso==0.8.3
path==16.7.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306; platform_system=="Windows"
pywin32-ctypes==0.2.2
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.3
traitlets==5.9.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.6
Werkzeug==3.0.1
whichcraft==0.6.1
wsproto==1.2.0
your-app==0.1.0
zipp==3.15.0
zope.event==4.5.0
zope.interface==5.4.0